rbac: inital RBAC implementation
This commit adds an initial implementation of RBAC for Observatorium
along with unit tests. RBAC rules can be read from a YAML file and look
a lot like Kubernetes RBAC rules. The only supported permissions are
`read` and `write` at the moment.

Once identification middleware is added to the stack, we can create e2e
tests for authorization.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>