internal/authentication: tenant ID in header
This commit adds a middleware that adds the tenant's ID to an HTTP
header that is forwarded to the metrics upstreams, e.g. so that Thanos
receive can forward the request to the correct hashring. It also
verifies that the middleware works as expected by explicitly filtering
tenants in the Thanos receive hashring configuration used in the
integration tests.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>