internal/authorization: create authz middleware
This commit implements and utilizes an authorization middleware to limit
API usage to subjects who have access to resources belonging to a
tenant, as specified in an RBAC configuration file.

At the heart of it, this commit simply ties the existing rbac.Authorizer
interface to a middleware. One implementation detail is that in order to
identify a subject, we need to find a username from the OIDC token. We
now allow the tenant to specify the field that should be used to
uniquely identify subjects, just as the Kubernetes apiserver does
with its `--oidc-username-claim` flag. The subject's name wil default to
the value of the `sub` field, but can be changed to, e.g., `email` in
order to have nicer usernames in RBAC role bindings.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>