.circleci: fix tagged master job
Currently, merging to master triggers the tagged-master job, because it
seems that CircleCI filters are ORed. This causes the Makefile to try to
tag a docker container with a trailing `:` since the Git tag is emtpy.
This commit changes the logic so that the tagged-master job is only
triggered on a valid SemVer tag.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>