internal/authentication: enable mTLS auth
Currently, the Observatorium API has overlapping authentication
concepts:
* the API has a global mTLS setting to validate client certs; and
* each tenant can authenticate clients via OIDC.

It is confusing and messy that these overlap. Identify should be
mediated entirely by the mechanism chosen by the tenant and the API's
TLS configuration should provide a secure transport.

This commit removes the global mTLS settings of the API and enables
mTLS-based authentication on a per-tenant basis.
This means that each tenant can specify a separate CA to authenticate
all requests for resources they own.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>