jsonnet: decouple deployment from secret
PR #87 was intended to decouple the deployment from a secret with a
magic, hard-coded name, but it obligated all consumers of this library
to render a secret and configmap using this jsonnet, rather than truly
decouple the deployment from the configuration files. This meant that
e.g. a custom resource for the Observatorium operator would need to have
the certificates in-line in order to render this manifest, which caused
two issues:
1. it put secrets in otherwise plain-text files; and
2. it prevented (re-)using secrets that may be created by another
service, e.g. cert-manager.

This PR fixes this by truly decoupling the deployment and the
configuration files. Now, the jsonnet accepts a secret name and keys to
specify where the API can find the certificates it needs.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>