go.mod: bump prom-label-proxy
This commit bumps the dependency on prom-label-proxy to include
https://github.com/prometheus-community/prom-label-proxy/pull/71.
As described in that PR, the reason for bumping this is to take
advantage of the updates to the PromQL parser. Without this update,
the prom-label-proxy, and thus the Observatorium API cannot enforce
labels on queries that use any new features, e.g. queries that include:
* the last_over_time, sgn, clamp functions, added in
  prometheus/prometheus#8457;
* the @ operator, added in
  prometheus/prometheus#8121; or
* the group aggregation operator, added in
  prometheus/prometheus#7480.

Currently, queries including these functions produce errors like:
`parse error: unknown function with name "group"`

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>