Introduce Plugable Authenticators
currently, only OIDC or mTLS based authentications are possible.
However, there are cases where custom authenticators are needed, such
as authenticators that might need to use private go modules, therefore,
this commit refactors the current implementation of the authentication
logic to make it pluggable in order to support new authenticators.