Introduce openshift authenticator plugin
create a new authenticator for openshift, this is basically copying
the logic of the openshift authenticator to a new openshift
authenticator
pluginn.

This commit also rework the way the routes for the authenticators are
setup, in a way it can be more generic to add more routes patterns for
different authenticators.