rules: add custom UnmarshalJSON function
Currently, when a group of rules is unmarshaled, the concrete type in Go
will never be a rules.RecordingRule or a rules.AlertingRule, although
those types are defined in the rules package. Instead those rules will
always be some map[string]interface{} type. This means that it's not
possible for users of the package to perform type assertions on the
rules that are fetched via the client, which is a serious annoyance for
anyone using the package.

This commit fixes this by providing a custom unmarshaling function.
Furthermore, in order to fix unmarshaling via both JSON and YAML, the
commit switches the YAML package to ghodss/yaml, which is a superset of
the current YAML package, go-yaml/yaml.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>