Enforce Labels on Returned Rules (#191)
* rules/spec.yaml: fix enum

Rules are either Recording Rules or Alerting Rules, but not both. This
commit fixes this fact in the OpenAPI specification.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* rules: add custom UnmarshalJSON function

Currently, when a group of rules is unmarshaled, the concrete type in Go
will never be a rules.RecordingRule or a rules.AlertingRule, although
those types are defined in the rules package. Instead those rules will
always be some map[string]interface{} type. This means that it's not
possible for users of the package to perform type assertions on the
rules that are fetched via the client, which is a serious annoyance for
anyone using the package.

This commit fixes this by providing a custom unmarshaling function.
Furthermore, in order to fix unmarshaling via both JSON and YAML, the
commit switches the YAML package to ghodss/yaml, which is a superset of
the current YAML package, go-yaml/yaml.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* rules/rules.go: regenerate

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* api: enforce tenant on generated time series

Currently, recording and alerting rules that are fetched from the new
rules/raw API will not have any tenant label attached to them. This
means that when the rules are configured on a Thanos Ruler, the time
series that are generated will not have a tenant label on them. This is
a big problem, because without this tenant label, the very tenant to
whom the rules belong cannot query for the time series, making them in
effect useless. This commit enforces that all rules returned by the
rules/raw endpoint have a tenant_id label attached to them.

Note: this is a distinct problem from ensuring that the rules written by
a tenant are guaranteed to only include data from that tenant.
Currently, the rules from a tenant can select data from any tenant. This
is a distinct problem and should absolutely be tackled in a follow up.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>