Stop OIDC flow when remote writing with bad auth header
When handling remote write requests, if the Authorization
header isn't present, we don't want to redirect users back
to the identity provider for login.

The reason behind this is that tenants can overload the
identity provider in case they have bad configuration.
We want to be friendly and help prevent this issue from
happening.

Signed-off-by: Douglas Camata <159076+douglascamata@users.noreply.github.com>