Change write access redirect protection approach
Each handler is responsible for providing a "write middleware injection entrypoint". And it is used to avoid redirect by the clients writing data.

At the same time, the code as adapted to pass all the middlewares to the subrouters when they are created instead of on the upper level routers. This is required because the protection has to be mounted before the OIDC authentication.

Signed-off-by: Douglas Camata <159076+douglascamata@users.noreply.github.com>