authentication: clean up naming (#327)
* authentication: clarify naming

This commit tries to clarify the naming of variables in the
authentication module. Right now, we are using a capitalized identifier,
namely `ProviderFactory` for a local variable, which is odd in Golang.
Futhermore, this name incidentally shadows the `ProviderFactory` type.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* authentication: use consistent naming

Right now, the `ProviderFactory` is used to create a variable called
`authenticator`, which is of the `Provider` type. I find the mixed usage
of `provider` and `authenticator` somewhat confusing and think it would
be nice to standardize this since they are already somewhat abstract
concepts.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* authentication: rename authenticator to provider

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>