Trying to address FIPS issue with the observatorium service
FIPS CVE-2023-3089 requires TLS 1.2 only for OCP 4.10 and 4.11 since the
TLS 1.3 ciphers have been submitted but not yet approved.  I think
Observatorium should not work on these FIPS enabled environments when
that CVE is fixed.  With this change I am hoping it would fix the
issue.  This is based on a slack thread conversation though

Signed-off-by: Gus Parvin <gparvin@redhat.com>